backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.2
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic[email]==2.9.2
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
bcrypt==4.2.0
python-multipart==0.0.9
slowapi==0.1.9
redis==5.1.0
celery==5.4.0
qrcode[pil]==7.4.2
stripe==10.12.0
boto3==1.35.0
firebase-admin==6.5.0
twilio==9.3.4
httpx==0.27.2
pytest==8.3.3
pytest-asyncio==0.23.8
